# swagger/openapi.yaml
openapi: "3.0.3"
info:
  version: "0.2.1"
  title: "Front-end Technical Exercise"
  contact:
    email: "allandemiranda@gmail.com"
servers:
  - url: https://localhost:5000/api
...
paths:
...
  /sale:
    post:
      tags:
        - "Sale"
      summary: "Create sale"
      requestBody:
        required: true
        description: ""
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: "string"
                  description: "Product ID"
                quantity:
                  type: "number"
                  format: "int"
                  description: "Quantity of products"
                total:
                  type: "number"
                  format: "float"
                  description: "Total of sale"
              required:
                - product_id
                - quantity
                - total
      responses:
        "201":
          description: "A sale object"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    items:
                      $ref: "#/components/schemas/Sale"
        "500":
          description: "Server error"
        default:
          description: "Unexpected error"
      security:
        - bearerAuth: []
  /sale/{sale_id}:
    get:
      tags:
        - "Sale"
      summary: "Get a specific sale"
      description: ""
      parameters:
        - name: sale_id
          in: path
          required: true
          description: The ID of the sale to return
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: "A sale"
          content:
            application/json:
              schema:
                type: object
                properties:
                  sale:
                    items:
                      $ref: "#/components/schemas/Sale"
        "500":
          description: "Server error"
        default:
          description: "Unexpected error"
      security:
        - bearerAuth: []
    put:
      tags:
        - "Sale"
      summary: "Modify a specific sale"
      description: ""
      parameters:
        - name: sale_id
          in: path
          required: true
          description: The ID of the sale to return
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        description: ""
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: "number"
                  format: "int"
                  description: "Quantity of products"
                total:
                  type: "number"
                  format: "float"
                  description: "Total of sale"
      responses:
        "200":
          description: "A sale"
          content:
            application/json:
              schema:
                type: object
                properties:
                  sale:
                    items:
                      $ref: "#/components/schemas/Sale"
        "500":
          description: "Server error"
        default:
          description: "Unexpected error"
      security:
        - bearerAuth: []
  /sales/:
    get:
      tags:
        - "Sale"
      summary: "Get sales"
      description: ""
      responses:
        "200":
          description: "The sales"
          content:
            application/json:
              schema:
                type: object
                properties:
                  sales:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sale"
        "500":
          description: "Server error"
        default:
          description: "Unexpected error"
      security:
        - bearerAuth: []
# Product route 
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int"
        name:
          type: "string"
          description: "Full name"
        phone:
          type: "string"
          description: "Phone number"
        user_name:
          type: "string"
          description: "Login name"
      xml:
        name: "User"
    Sale:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int"
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: "integer"
          format: "int"
          description: "Quantity of products bought"
        total:
          type: "number"
          format: "float"
          description: "Total bought"
        date:
          type: "string"
          description: "Date bought"
        user:
          $ref: "#/components/schemas/User"
      xml:
        name: "Sale"
    Product:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int"
        name:
          type: "string"
          description: "name of product"
        value:
          type: "number"
          format: "float"
          description: "Value of product"
      xml:
        name: "Product"